Make functions not depend on objects in the workspace.
Some of the functions assumed that n.ctrls, n.cases, xseq and x.stepsize
were available in the global workspace. By storing these quantities in the
object produced by Wdensities.unadjusted(), we can make the code safer and
more consistent.

This changes the signature of Wdensities.fromraw() to not require the y
variable.