Fix the horizontal range of plotcumfreqs().
The previous code used W from the working space, so it would fail if this did
not exist. Setting it to use the W stored in the densities object caused the
horizontal range to be too wide (from -25 to 25 by default).

The new approach finds the range in which both curves are sufficiently far
away from 0 or 1, and thus the plot only includes the most interesting part
of the graph.

This fixes the last remaining note raised by R CMD check.