Further simplification to the interface of Wdensities().
This allows to avoid the computation of the weights of evidence, which can be
done inside Wdensities() if the posterior and prior probabilities are passed
as arguments.

Also, storing those probabilities in the densities object returned will allow
writint a summary method that reports all crude and model-based performance
measures.