Speed up Wdensities.crude().
By using densities() directly rather than fsmooth(), the computation of
densities in cases and controls can be sped up quite a lot, especially for
large datasets. There are no perceivable differences in the plots nor in
the quantities computed by summary. The mean relative differences are about
1e-5 or less.

On an example with 45000 observations (44361 controls, 639 cases) it takes
now 11 seconds instead of 40.