Fix error in computing mixture densities when there is only one component.
It may happen that Wdensities() decides that we should fit 2 components when
computing densities, but either all cases or all controls fall in the same
mixture component. If this is the case, the following error was thrown:

   Density with 2 mixture components chosen by BIC
   Error in density.default(W[mixcomponent == 1], bw = "SJ", n = num.xseq,  :
     need at least 2 points to select a bandwidth automatically

This changes densities.mix() to check for this situation and compute only one
density if in that subset (cases or controls) all observations fall in the
same mixture component.